+++ objectbox/CHANGELOG.md
@@ -13,6 +13,8 @@
 * Add `StorageException` which is a `ObjectBoxException` with an `errorCode` (a `OBX_ERROR` code).
 * Throw `DbFullException` instead of `ObjectBoxException` with message `10101 Could not put` (error 
   code `OBX_ERROR_DB_FULL`).
+* Change `Query.findUnique()` to throw `NonUniqueResultException` instead of 
+  `UniqueViolationException` if there is more than one result.
 
 ## 1.6.2 (2022-08-24)